Content types: extract labels auto-fill action into shared util
Lifts the LabelsActionsEdit component and the phantom Field that wraps it
out of taxonomies/fields/labels.tsx and into a shared
createLabelsActionsField factory in src/utils/labels.tsx. The factory
takes the per-content-type label keys, deriveLabels function, and help
text, so post types can reuse the same Auto-fill / Clear UI.

Behaviour-preserving for taxonomies.
