+++ packages/content-types/src/taxonomies/fields/labels.tsx
@@ -1,18 +1,18 @@
 /**
  * WordPress dependencies
  */
-import { Button } from '@wordpress/components';
-import type { DataFormControlProps, Field, Form } from '@wordpress/dataviews';
+import type { Form } from '@wordpress/dataviews';
 import { __ } from '@wordpress/i18n';
-import { trash } from '@wordpress/icons';
-
-import { Stack, Text } from '@wordpress/ui';
 
 /**
  * Internal dependencies
  */
 import { createLabelField } from '../../utils/fields';
-import type { StoredLabels, TaxonomyFormData } from '../types';
+import {
+	createLabelsActionsField,
+	LABELS_ACTIONS_FIELD_ID,
+} from '../../utils/labels';
+import type { TaxonomyFormData } from '../types';
 import { deriveLabels, STRING_LABEL_KEYS } from '../utils';
 
 export const menuNameField = createLabelField( 'menu_name', __( 'Menu name' ), {
@@ -126,90 +126,19 @@ export const chooseFromMostUsedField = createLabelField(
 	}
 );
 
-function LabelsActionsEdit( {
-	data,
-	onChange,
-}: DataFormControlProps< TaxonomyFormData > ) {
-	const plural = data.title.raw.trim();
-	const singular = data.config.labels.singular_name.trim();
-	const canAutoFill = !! plural.length && !! singular.length;
-	const hasOverrides = STRING_LABEL_KEYS.some(
-		( key ) =>
-			key !== 'singular_name' &&
-			( data.config.labels[ key ] ?? '' ) !== ''
-	);
-	return (
-		<Stack direction="column" gap="md">
-			<Text variant="body-md" className="taxonomy-form__help-text">
-				{ __(
-					'Override the text WordPress shows in admin lists, menus, and forms. Auto-fill replaces every label below with values derived from the current plural and singular names — including any you have already customized. Clearing removes all overrides so WordPress falls back to its defaults. If you rename the taxonomy after auto-filling, click Auto-fill again to keep them in sync.'
-				) }
-			</Text>
-			<Stack direction="row" justify="flex-end" gap="sm">
-				<Button
-					__next40pxDefaultSize
-					variant="secondary"
-					size="compact"
-					accessibleWhenDisabled
-					disabled={ ! canAutoFill }
-					onClick={ () =>
-						onChange( {
-							config: {
-								...data.config,
-								labels: {
-									...data.config.labels,
-									...deriveLabels( plural, singular ),
-								},
-							},
-						} )
-					}
-				>
-					{ __( 'Auto-fill labels' ) }
-				</Button>
-				<Button
-					__next40pxDefaultSize
-					size="compact"
-					icon={ trash }
-					isDestructive
-					label={ __( 'Clear labels' ) }
-					accessibleWhenDisabled
-					disabled={ ! hasOverrides }
-					onClick={ () => {
-						const cleared: StoredLabels = {
-							singular_name: data.config.labels.singular_name,
-						};
-						for ( const key of STRING_LABEL_KEYS ) {
-							if ( key !== 'singular_name' ) {
-								cleared[ key ] = '';
-							}
-						}
-						onChange( {
-							config: { ...data.config, labels: cleared },
-						} );
-					} }
-				/>
-			</Stack>
-		</Stack>
-	);
-}
-
-// TODO: Replace this phantom field once DataForm supports per-card header
-// actions or arbitrary React among children. We register a no-value Field
-// solely to render the Labels card's description and action buttons,
-// suppressing its label via labelPosition: 'none'.
-export const labelsActionsField: Field< TaxonomyFormData > = {
-	id: '__labels_actions',
-	label: '',
-	getValue: () => '',
-	setValue: () => ( {} ),
-	Edit: LabelsActionsEdit,
-	enableSorting: false,
-	filterBy: false,
-};
+export const labelsActionsField = createLabelsActionsField< TaxonomyFormData >(
+	{
+		labelKeys: STRING_LABEL_KEYS,
+		deriveLabels,
+		helpText: __(
+			'Override the text WordPress shows in admin lists, menus, and forms. Auto-fill replaces every label below with values derived from the current plural and singular names — including any you have already customized. Clearing removes all overrides so WordPress falls back to its defaults. If you rename the taxonomy after auto-filling, click Auto-fill again to keep them in sync.'
+		),
+	}
+);
 
 export const labelsFormFields: Form[ 'fields' ] = [
 	{
-		id: '__labels_actions',
+		id: LABELS_ACTIONS_FIELD_ID,
 		layout: { type: 'regular', labelPosition: 'none' },
 	},
 	// singular_name lives in the General card, so exclude it here.
